+++ src/main/java/com/pax/market/api/sdk/java/api/terminal/dto/TerminalReplacementUpdateRequest.java
@@ -12,20 +12,11 @@
 package com.pax.market.api.sdk.java.api.terminal.dto;
 
 
-import java.io.Serializable;
+public class TerminalReplacementUpdateRequest extends TerminalConfigUpdateRequest {
 
-/**
- * @Description
- * @Author: Shawn
- * @Date: 2020/4/24
- */
-
-public class TerminalConfigDTO implements Serializable {
-
-    private static final long serialVersionUID = 6509938259306387562L;
+    private static final long serialVersionUID = 1L;
     private Boolean allowReplacement;
-    private Boolean automaticTimezoneEnable;
-    private String timeZone;
+
 
     public Boolean getAllowReplacement() {
         return allowReplacement;
@@ -35,19 +26,4 @@ public class TerminalConfigDTO implements Serializable {
         this.allowReplacement = allowReplacement;
     }
 
-    public Boolean getAutomaticTimezoneEnable() {
-        return automaticTimezoneEnable;
-    }
-
-    public void setAutomaticTimezoneEnable(Boolean automaticTimezoneEnable) {
-        this.automaticTimezoneEnable = automaticTimezoneEnable;
-    }
-
-    public String getTimeZone() {
-        return timeZone;
-    }
-
-    public void setTimeZone(String timeZone) {
-        this.timeZone = timeZone;
-    }
 }