+++ src/main/java/com/pax/market/api/sdk/java/api/terminal/dto/TerminalTimeZoneUpdateRequest.java
@@ -12,35 +12,19 @@
 package com.pax.market.api.sdk.java.api.terminal.dto;
 
 
-import java.io.Serializable;
+public class TerminalTimeZoneUpdateRequest extends TerminalConfigUpdateRequest {
 
-/**
- * @Description
- * @Author: Shawn
- * @Date: 2020/4/24
- */
+    private static final long serialVersionUID = 1L;
 
-public class TerminalConfigDTO implements Serializable {
-
-    private static final long serialVersionUID = 6509938259306387562L;
-    private Boolean allowReplacement;
-    private Boolean automaticTimezoneEnable;
+    private Boolean automaticTimeZoneEnable;
     private String timeZone;
 
-    public Boolean getAllowReplacement() {
-        return allowReplacement;
+    public Boolean getAutomaticTimeZoneEnable() {
+        return automaticTimeZoneEnable;
     }
 
-    public void setAllowReplacement(Boolean allowReplacement) {
-        this.allowReplacement = allowReplacement;
-    }
-
-    public Boolean getAutomaticTimezoneEnable() {
-        return automaticTimezoneEnable;
-    }
-
-    public void setAutomaticTimezoneEnable(Boolean automaticTimezoneEnable) {
-        this.automaticTimezoneEnable = automaticTimezoneEnable;
+    public void setAutomaticTimeZoneEnable(Boolean automaticTimeZoneEnable) {
+        this.automaticTimeZoneEnable = automaticTimeZoneEnable;
     }
 
     public String getTimeZone() {