+++ src/main/java/com/pax/market/api/sdk/java/api/terminalGeoFenceWhiteList/dto/TerminalGeoFenceWhiteListPageResponse.java
@@ -1,0 +1,10 @@
+package com.pax.market.api.sdk.java.api.terminalGeoFenceWhiteList.dto;
+
+import com.pax.market.api.sdk.java.api.base.dto.PageResponse;
+
+
+public class TerminalGeoFenceWhiteListPageResponse extends PageResponse<TerminalGeoFenceWhiteListDTO> {
+    private static final long serialVersionUID = 8001293793762041406L;
+
+
+}