+++ src/test/java/com/pax/market/api/sdk/java/api/test/TerminalApiTest.java
@@ -29,6 +29,7 @@ import com.pax.market.api.sdk.java.api.terminal.dto.TerminalUpdateRequest;
 
 import java.util.HashSet;
 import java.util.Set;
+import java.util.TimeZone;
 
 
 /**
@@ -165,11 +166,23 @@ public class TerminalApiTest {
 	}
 
 	@Test
-	public void testUpdateTerminalConfig(){
+	public void testUpdateTerminalReplacementConfig(){
 		Long terminalId = 909744L;
-		TerminalConfigUpdateRequest terminalConfigUpdateRequest = new TerminalConfigUpdateRequest();
-		terminalConfigUpdateRequest.setAllowReplacement(true);
-	    Result<String> result = terminalApi.updateTerminalConfig(terminalId,terminalConfigUpdateRequest);
+		TerminalReplacementUpdateRequest terminalReplacementUpdateRequest = new TerminalReplacementUpdateRequest();
+		terminalReplacementUpdateRequest.setAllowReplacement(true);
+	    Result<String> result = terminalApi.updateTerminalConfig(terminalId, terminalReplacementUpdateRequest);
+		logger.debug("Result of update Terminal Config: {}",result.toString());
+		Assert.assertTrue(result.getBusinessCode() == 0);
+
+	}
+
+	@Test
+	public void testUpdateTerminalTimeZoneConfig(){
+		Long terminalId = 1547947831459887L;
+		TerminalTimeZoneUpdateRequest terminalTimeZoneUpdateRequest = new TerminalTimeZoneUpdateRequest();
+		terminalTimeZoneUpdateRequest.setAutomaticTimeZoneEnable(false);
+		terminalTimeZoneUpdateRequest.setTimeZone(TimeZone.getDefault().getID());
+		Result<String> result = terminalApi.updateTerminalConfig(terminalId, terminalTimeZoneUpdateRequest);
 		logger.debug("Result of update Terminal Config: {}",result.toString());
 		Assert.assertTrue(result.getBusinessCode() == 0);